adding first room guide line with boss animation but still wip
