+++ client/src/components/game/gameCommands/minigame/prompt/Prompt.css
@@ -1,0 +1,12 @@
+.prompt {
+  flex-direction: column;
+}
+
+.prompt-button {
+  margin-top: 2rem;
+  padding: 0.5rem;
+  background-color: white;
+  color: rgb(209, 41, 7);
+  font-weight: bold;
+  border-radius: 5px;
+}
